Upct

+++ app/Http/Controllers/IntrestController.php
@@ -23,7 +23,7 @@ class IntrestController extends Controller
 
             return $this->returnData($intrests,'operation completed successfully');
         } catch (\Exception $ex) {
-            return $this->returnError($ex->getCode(),$ex->getMessage());
+            return $this->returnError("500",$ex->getMessage());
         }
     }
 
@@ -43,7 +43,7 @@ class IntrestController extends Controller
             return $this->returnData($intrest,'operation completed successfully');
         } catch (\Exception $ex) {
             DB::rollback();
-            return $this->returnError($ex->getCode(), 'Please try again later');
+            return $this->returnError("500", 'Please try again later');
         }
     }
 
@@ -81,16 +81,18 @@ class IntrestController extends Controller
         try {
             DB::beginTransaction();
             $profile_student=auth()->user()->profile_student()->first();
-            $intrest=$profile_student->intrests()->find($id);
-            if(!$intrest)
-                return $this->returnError("404", 'not found');
-            $intrest->delete();
+            if($profile_student) {
+                $intrest = $profile_student->intrests()->find($id);
+                if (!$intrest)
+                    return $this->returnError("404", 'not found');
+                $intrest->delete();
+            }
 
             DB::commit();
             return $this->returnSuccessMessage('operation completed successfully');
         } catch (\Exception $ex) {
             DB::rollback();
-            return $this->returnError($ex->getCode(), 'Please try again later');
+            return $this->returnError("500", 'Please try again later');
         }
     }
 }
